Created initial skeleton for project. Not much of a skeleton, but it does demonstrate a few things. For one, it is officially a proof of concept for the issue of displaying data. So far, the data that is collected should be mapped to a GraphData object, which will handle not only timestamping that data, but as well keeping track of the data itself. This is extremely easy to have in a database and in memory, and doubly so for manipulating at runtime (Observables and Streams anyone?).
As well, I defined a "Provider" interface but I am not sure if I want to
go through with it. A Provider would essentially provide the user with
data. This however may not work anyway if the modular approach is
needed.

As well, I have a decent graph up and running with randomly generated
stubbed data. So far it does a decent job of testing, and I can say that
chart.js definitely is a very handy library to use.