Kevin development (#61)
* Committing to my own branch to merge with Development for Tyler's updates

* Updated Database manager

Added description update, insert, and get functions

* updated databasemananger

* Changed path to properties

Hard-coded path to properties for a good local connection

* Keeping changes to DatabaseManager

* Added the working DatabaseManager back into the package

* name -> dataName fix

* Committing to Brandon-Development to continue working from lab

* Toyed with DatabaseManager (then undid all my changes), tried managing the user table with UserManager

* Beginning Admin

* Unsure of Dan's progress, pushing ahead with admin functions and GUI

* More work on admin functionality

* More Admin progress

No functionality, add entry button not connected to a method. Beginning
work on Delete Data tab

* Admin work

* Updated styles for Dashboard GUI

* fixed the limits

* Group development (#48)

Group development

* changed the export tab to a export button that blends in with the other
tabs.

* Added double Y-Axis and default to Temperature of Air and Water

* Began revising so that the data parameter chosen can  is based on it's actual id

* Refactored DataReceiver a bit so that it maintains a single PARAMETER_MAP, which keeps track of DataParameters, which is an object that keeps track of the name, id, unit, etc., for a given parameter

* Updated .gitignore

Moving the GitHub folder to My Documents caused a small issue with
thumbs.db, this resolves that issue

* Started playing around with Highchartjs in order to eventually make the
graph zoomable with a double axis for 2 data points

* Added a way to get just raw data from the Data returned from DataReceiver for people who would prefer to use it over Observables

* Styled the Highcharts chart to match the styling of the rest of the page,
and and played around trying to get real data in there

* Updated the chart so that it pulls data from the request attributes, so Kevin can focus purely on front end and not backend.

* Took the Highchart graph that Louis was able to make to generate real data
and styled it to zoom, resize, and a couple other minor changes.

* Created body of admin AJAX/jQuery insertion template

Using the example I followed online, I typed up basic insertion and
deletion functions to test. Now I am translating it to our project,
changing variables as necessary. Testing with our JSON's comes next.

* Made it so there is a dedicated form for the graph and table tab
also made it so that whichever tab was submitted from is the tab that 
loads after submission

* Change in properties: WaterQualityServer now WaterQuality

I changed the name of the project per Dr. Jones' request. It is no
longer WaterQualityServer. The properties files need to be kept to keep
the change.

Also committing work on the javascript files to implement AJAX requests
on the dashboard.

* Just changed it back so that table still generates when  graph is made

* just fixing stuff

* L.J: Fixed some issues for Kevin

* added the capability for double axes. Made it display with the color of
the lines and dynamically places names of data based on whatever is checked