Dan development - Renaming and Deletion of Admin Pages (#69)
* Admin Pages minor fixes

Minor changes on Dan's and began some changes on Brandon's

* Styling and Design

Made changes to the styling and design of Brandon's version of the admin
pages.  Also did some renaming of CSS classes.

* Dan-Development wk4 (#67)

* Admin Pages minor fixes

Minor changes on Dan's and began some changes on Brandon's

* Styling and Design

Made changes to the styling and design of Brandon's version of the admin
pages.  Also did some renaming of CSS classes.

* conflict fix

* Renaming and Deletions

Removed old versions of admin page.