Kevin development (#93)
* Made some style changes to the table generated for the table tab such as
changing the header color and centering the table in the tab.
Also added the ability to click on the table and a popup appears of the
table making the table easier to view. For this I added the popup.css file
which is the style sheet for the popup. Added a modal box for the popup
table, and openPop function for opening the popup

* Added the exportTable() function to export table data to a csv file
also made some style changes to the popup table

* Added an exportGraph() function so that the graph can now be exported
along with the table.