Bayesian loading (#141)
* Made some DataTables changes as well... tedious as hell, definitely 3 hours worth of work doing next to nothing. Web Development is so boring...

* Added a loader to the bayesian tab.

* Fixed the issue with our data being a few hours behind the actual data.
Also was trying to fix the dataTables since it auto sorts the datetime
data in an incorrect way.

* Style Changes to make things work better

* Got rid of my temporary fix for our data being behind since there was a
different issue causing it to not work properly. Fixed the datatables
sorting by adding a hidden date formatted for the datatables to be able
to sort by date correctly

* The datatable was still ordering slightly wrong so I fixed it and orders
correctly this time.

* Made some progress on fixing issues caused by Loius's new code. Fixed 
the fact that if a parameter was selected in the graph it would show up
on the table.

* Fixes for Graph/table data loading

Both manual/sensor data will now display descriptions. Completely fixed
the issue with datatable sorting icorrectly. Fixed the fact that nothing
would display if only manual data was selected for the table.